+++ packages/cdktf/lib/terraform-stack.ts
@@ -7,6 +7,7 @@ import { deepMerge } from './util';
 
 export class TerraformStack extends Construct {
   public readonly artifactFile: string;
+  private readonly rawOverrides: any = {}
 
   constructor(scope: Construct, id: string) {
     super(scope, id);
@@ -14,12 +15,35 @@ export class TerraformStack extends Construct {
     this.artifactFile = `${Node.of(this).uniqueId}.tf.json`;
   }
 
+  public addOverride(path: string, value: any) {
+    const parts = path.split('.');
+    let curr: any = this.rawOverrides;
+
+    while (parts.length > 1) {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      const key = parts.shift()!;
+
+      // if we can't recurse further or the previous value is not an
+      // object overwrite it with an object.
+      const isObject = curr[key] != null && typeof(curr[key]) === 'object' && !Array.isArray(curr[key]);
+      if (!isObject) {
+        curr[key] = {};
+      }
+
+      curr = curr[key];
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const lastKey = parts.shift()!;
+    curr[lastKey] = value;
+  }
+
   public toTerraform(): any {
-    const tf = { };
+    const tf = {};
 
     const visit = (node: IConstruct) => {
       if (node instanceof TerraformElement) {
-        deepMerge(tf, node.toTerraform());
+        deepMerge(tf,  node.toTerraform());
       }
 
       for (const child of Node.of(node).children) {
@@ -29,7 +53,9 @@ export class TerraformStack extends Construct {
 
     visit(this);
 
-    return resolve(this, tf)
+    deepMerge(tf, this.rawOverrides);
+
+    return resolve(this, tf);
   }
 
   public onSynthesize(session: ISynthesisSession) {